gpu/requirements.txt
torch>=2.0
torchvision>=0.15
pyzmq>=25.0
PyYAML>=6.0
numpy>=1.24
scipy>=1.10
wandb>=0.15
